+++ package.json
@@ -6,8 +6,8 @@
     "@rollup/plugin-commonjs": "^11.0.0",
     "@rollup/plugin-node-resolve": "^6.0.0",
     "@semantic-release/commit-analyzer": "^6.3.1",
-    "@types/fetch-mock": "^7.3.1",
-    "@types/jasmine": "^3.4.0",
+    "@types/fetch-mock": "^7.3.2",
+    "@types/jasmine": "^3.5.10",
     "@types/node": "^10.17.0",
     "acetate": "^2.0.8",
     "acetate-cli": "^1.0.1",
@@ -23,7 +23,7 @@
     "cz-lerna-changelog": "^2.0.2",
     "date-fns": "^1.29.0",
     "event-stream": "3.3.4",
-    "fetch-mock": "^7.7.0",
+    "fetch-mock": "^7.7.3",
     "gh-pages": "^1.1.0",
     "gh-release": "^3.4.0",
     "handlebars": "^4.5.1",
